# argos-service-api/api.yml
...
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Argos Service
  license:
    name: Apache License 2.0
servers:
  - url: /api
security:
  - bearerAuth: []
paths:
...
  /supplychain/{supplyChainId}/account:
    get:
      summary: searches for service accounts and personal accounts
      operationId: searchAccounts
      security:
        - bearerAuth: []
      tags:
        - searchAccount
      parameters:
        - name: supplyChainId
          in: path
          description: supply chain id
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
        - name: name
          in: query
          required: true
          schema:
            type: string
            maxLength: 1024
        - name: accountType
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/AccountType"
      responses:
        '200':
          description: array of AccountInfo
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountInfo"
        '500':
          $ref: '#/components/responses/Error'
...
  /personalaccount/me/key:
    post:
      summary: creates a new active public and private key pair
      operationId: createKey
      tags:
        - personal account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeyPair"
      responses:
        '204':
          description: no content
        '404':
          description: personal account not found
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/Error'

    get:
      summary: gets the currently active a key pair
      operationId: getKeyPair
      tags:
        - personal account
      responses:
        '200':
          description: Key pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeyPair"
        '404':
          description: personal account not found
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/Error'
  /personalaccount/me:
    get:
      summary: returns the personal account profile of the authenticated personal account based on the bearer token
      operationId: getPersonalAccountOfAuthenticatedUser
      tags:
        - personal account
      responses:
        '200':
          description: Personal Account Profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '404':
          description: personal account not found
        '500':
          $ref: '#/components/responses/Error'
...
  /personalaccount/me/refresh:
    get:
      summary: refresh token
      operationId: refreshToken
      tags:
        - personal account
      responses:
        '200':
          description: new token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '500':
          $ref: '#/components/responses/Error'
  /personalaccount:
    get:
      summary: search
      operationId: searchPersonalAccounts
      tags:
        - personal account
      parameters:
        - name: roleName
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/NameString"
        - name: localPermissionsLabelId
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/UUID"
        - name: name
          in: query
          required: false
          schema:
            type: string
            maxLength: 1024
        - name: activeKeyIds
          in: query
          required: false
          explode: false
          schema:
            type: array
            maxItems: 256
            items:
              $ref: "#/components/schemas/SHA256Hash"
        - name: inactiveKeyIds
          in: query
          required: false
          explode: false
          schema:
            type: array
            maxItems: 256
            items:
              $ref: "#/components/schemas/SHA256Hash"
      responses:
        '200':
          description: list of personal accounts without roles property
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PersonalAccount"
  /personalaccount/{accountId}:
    get:
      summary: get personal account by id
      operationId: getPersonalAccountById
      tags:
        - personal account
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        '200':
          description: personal account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonalAccount"
        '500':
          $ref: '#/components/responses/Error'
  /personalaccount/{accountId}/key:
    get:
      summary: get active key by personal account id
      operationId: getPersonalAccountKeyById
      tags:
        - personal account
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        '200':
          description: key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicKey"
        '404':
          description: no active key exists
        '500':
          $ref: '#/components/responses/Error'
  /personalaccount/{accountId}/role:
    put:
      summary: update personal account roles by id
      operationId: updatePersonalAccountRolesById
      tags:
        - personal account
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: role names
              type: array
              maxItems: 20
              items:
                $ref: "#/components/schemas/NameString"
      responses:
        '200':
          description: updated personal account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonalAccount"
        '500':
          $ref: '#/components/responses/Error'
  /personalaccount/{accountId}/localpermission/{labelId}:
    get:
      summary: returns the local permissions of the personalaccount and label
      operationId: getLocalPermissionsForLabel
      tags:
        - personal account
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
        - name: labelId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        '200':
          description: local permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocalPermissions"
        '204':
          description: no local permissions
    put:
      summary: updates permissions of the personalaccount and label when the permissions list is empty the permissions are removed
      operationId: updateLocalPermissionsForLabel
      tags:
        - personal account
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
        - name: labelId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: local permission
              type: array
              maxItems: 256
              items:
                $ref: "#/components/schemas/Permission"
      responses:
        '200':
          description: local permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocalPermissions"
        '204':
          description: local permissions removed
  /personalaccount/{accountId}/localpermission:
    get:
      summary: returns all the local permissions of the personalaccount
      operationId: getAllLocalPermissions
      tags:
        - personal account
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        '200':
          description: all local permissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LocalPermissions"
        '500':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
  responses:
    ValidationError:
      description: validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
    Error:
      description: server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    SHA256Hash:
      type: string
      minLength: 64
      maxLength: 64
      pattern: '^[0-9a-f]*$'
    UUID:
      type: string
      description: an Argos Service identifier
      minLength: 36
      maxLength: 36
      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}?$'
...
    NameString:
      type: string
      minLength: 1
      maxLength: 63
      pattern: '^([a-z]|[a-z][a-z0-9-]*[a-z0-9])?$'
    ByteString:
      type: string
      maxLength: 4096
      format: byte
...
    PublicKey:
      type: object
      required:
        - keyId
        - publicKey
      properties:
        keyId:
          $ref: '#/components/schemas/KeyId'
        publicKey:
          $ref: '#/components/schemas/ByteString'
    KeyPair:
      type: object
      required:
        - keyId
        - publicKey
        - encryptedPrivateKey
      properties:
        keyId:
          $ref: '#/components/schemas/KeyId'
        publicKey:
          $ref: '#/components/schemas/ByteString'
        encryptedPrivateKey:
          $ref: '#/components/schemas/ByteString'
...
    KeyId:
      $ref: "#/components/schemas/SHA256Hash"
...
    PersonalAccount:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
          readOnly: true
        name:
          type: string
          maxLength: 1024
        roles:
          type: array
          maxItems: 20
          items:
            $ref: "#/components/schemas/Role"
    Profile:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
          readOnly: true
        name:
          type: string
          maxLength: 1024
        email:
          type: string
          maxLength: 1024
        roles:
          type: array
          maxItems: 20
          items:
            $ref: "#/components/schemas/Role"
...
    Role:
      type: object
      required:
        - name
      properties:
        id:
          $ref: "#/components/schemas/UUID"
          readOnly: true
        name:
          $ref: "#/components/schemas/NameString"
        permissions:
          type: array
          minItems: 1
          maxItems: 20
          items:
            $ref: "#/components/schemas/Permission"
    Permission:
      enum:
        - READ
        - TREE_EDIT
        - LOCAL_PERMISSION_EDIT
        - ASSIGN_ROLE
        - LINK_ADD
        - LAYOUT_ADD
        - VERIFY
        - RELEASE
        - PERSONAL_ACCOUNT_READ
        - SERVICE_ACCOUNT_EDIT
    LocalPermissions:
      type: object
      required:
        - labelId
        - permissions
      properties:
        labelId:
          $ref: "#/components/schemas/UUID"
        permissions:
          type: array
          maxItems: 20
          items:
            $ref: "#/components/schemas/Permission"
...
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    ValidationError:
      type: object
      required:
        - messages
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ValidationMessage"
    ValidationMessage:
      type: object
      required:
        - message
        - type
      properties:
        field:
          type: string
        type:
          enum:
            - DATA_INPUT
            - MODEL_CONSISTENCY
        message:
          type: string
    AccountType:
      enum:
        - SERVICE_ACCOUNT
        - PERSONAL_ACCOUNT
...
    AccountInfo:
      type: object
      properties:
        accountId:
          $ref: "#/components/schemas/UUID"
        name:
          type: string
          maxLength: 1024
        path:
          type: string
        accountType:
          $ref: '#/components/schemas/AccountType'
...
    Token:
      type: object
      required:
        - token
      properties:
        token:
          type: string